Move curve helpers to their own file
